requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
openai==1.3.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0